Move ERC721 tests to test-extras.
These can fail nondeterministically and should should not be run as part
of the usual CI test suite.
